backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
jinja2>=3.1.4
pydantic>=2.10.0
python-dotenv>=1.0.1
email-validator>=2.0.0
supabase>=2.0.0
httpx>=0.27.0
faster-whisper>=1.0.0
websockets